requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
autoflake==2.3.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cryptography==42.0.0
cssselect2==0.7.0
distlib==0.3.8
Django==5.0.1
django-admin-interface==0.28.6
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.2
django-js-asset==2.2.0
django-tailwind==3.8.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.13.1
html5lib==1.1
idna==3.6
isort==5.13.2
lxml==5.1.0
Markdown==3.6
mysqlclient==2.2.1
numpy==1.26.3
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
pyflakes==3.2.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.0
pypdf==4.0.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.25.1
webencodings==0.5.1
xhtml2pdf==0.2.14
XlsxWriter==3.1.9
